refactor keyboards to domain objects
