+++ SharpBoard.Desktop/View/CurrentKeyControl.xaml.cs
@@ -1,7 +1,7 @@
 ﻿using Injecter;
 using MediatR;
 using SharpBoard.Desktop.Extensions;
-using SharpBoard.Domain.Keyboards;
+using SharpBoard.Domain;
 using SharpBoard.KeysApi;
 using System.ComponentModel;
 using System.Runtime.CompilerServices;
@@ -21,11 +21,11 @@ namespace SharpBoard.Desktop.View
             set => SetAndNotifyProperty(ref _tesoroLedId, value);
         }
 
-        private KeyboardKind _currentKeyboard;
+        private KeyboardKind? _currentKeyboard;
 
         public KeyboardKind CurrentKeyboard
         {
-            get => _currentKeyboard;
+            get => _currentKeyboard ?? KeyboardKind.None;
             set => SetAndNotifyProperty(ref _currentKeyboard, value);
         }
