+++ SharpBoard.Desktop/View/MainWindow.xaml.cs
@@ -1,6 +1,5 @@
 ﻿using MoreLinq;
-using SharpBoard.Domain.Keyboards;
-using System;
+using SharpBoard.Domain;
 using System.Windows;
 using Xceed.Wpf.AvalonDock.Controls;
 
@@ -13,7 +12,7 @@ namespace SharpBoard.Desktop.View
             Loaded += OnLoaded;
             InitializeComponent();
 
-            KeyboardSelector.ItemsSource = Enum.GetValues<KeyboardKind>();
+            KeyboardSelector.ItemsSource = KeyboardKind.List;
         }
 
         private void OnLoaded(object sender, RoutedEventArgs e) =>
