+++ SharpBoard/CoreInjector.cs
@@ -1,7 +1,5 @@
 ﻿using MediatR;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.DependencyInjection.Extensions;
-using SharpBoard.Domain.Keyboards;
 
 namespace SharpBoard
 {
@@ -9,10 +7,6 @@ namespace SharpBoard
     {
         public static IServiceCollection AddCore(this IServiceCollection services)
         {
-            services.TryAddTransient<TesoroKeyboard>();
-            services.TryAddTransient<RedragonKeyboard>();
-            services.TryAddTransient<IKeyboardFactory, KeyboardFactory>();
-
             services.AddMediatR(typeof(CoreInjector).Assembly);
 
             return services;
