--- SharpBoard/Domain/Keyboards/IKeyboard.cs
@@ -1,10 +1,0 @@
-﻿using System.Threading;
-using System.Threading.Tasks;
-
-namespace SharpBoard.Domain.Keyboards
-{
-    public interface IKeyboard
-    {
-        Task SetColorValue(ColorRgb256 color, int keyId, CancellationToken cancellationToken = default);
-    }
-}
