--- SharpBoard/Domain/Keyboards/IKeyboardFactory.cs
@@ -1,7 +1,0 @@
-﻿namespace SharpBoard.Domain.Keyboards
-{
-    public interface IKeyboardFactory
-    {
-        IKeyboard CreateKeyboard(KeyboardKind keyboard);
-    }
-}
