--- SharpBoard/Domain/Keyboards/KeyboardKind.cs
@@ -1,8 +1,0 @@
-﻿namespace SharpBoard.Domain.Keyboards
-{
-    public enum KeyboardKind
-    {
-        Tesoro,
-        Redragon,
-    }
-}
