+++ SharpBoard/KeysApi/SetKeyColor.cs
@@ -1,6 +1,5 @@
 ﻿using MediatR;
 using SharpBoard.Domain;
-using SharpBoard.Domain.Keyboards;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -11,15 +10,9 @@ namespace SharpBoard.KeysApi
     {
         public sealed class Handler : IRequestHandler<SetKeyColor, Unit>
         {
-            private readonly IKeyboardFactory _factory;
-
-            public Handler(IKeyboardFactory factory) => _factory = factory;
-
             public async Task<Unit> Handle(SetKeyColor request, CancellationToken cancellationToken)
             {
-                var (color, keyId, keyboardKind) = request;
-
-                var keyboard = _factory.CreateKeyboard(keyboardKind);
+                var (color, keyId, keyboard) = request;
 
                 await keyboard.SetColorValue(color, keyId, cancellationToken);
